add servomotor module and refactor modules
